+++ doc/common/conf.py
@@ -33,10 +33,6 @@ TOP_DIR = os.path.abspath(os.path.join('..', '..'))
 sys.path.insert(0, os.path.join(TOP_DIR, 'src'))
 sys.path.append(os.path.abspath('../common/extensions'))
 
-from pybricks.hubs import EV3Brick  # noqa E402
-from pybricks.media.ev3dev import Image  # noqa E402
-from pybricks._common import Speaker  # noqa E402
-
 # ON_RTD is whether we are on readthedocs.org
 # this line of code grabbed from docs.readthedocs.org
 ON_RTD = os.environ.get('READTHEDOCS', None) == 'True'
