+++ src/pybricks/_common.pyi
@@ -5,7 +5,7 @@ from typing import Collection, Iterable, Optional, Tuple, Union, overload
 
 from .geometry import Axis, Matrix, vector
 from .parameters import Button, Color, Direction, Side, Stop, Port
-from .media.ev3dev import SoundFile
+
 
 class DCMotor:
     def __init__(
@@ -83,16 +83,6 @@ class Motor(DCMotor):
 class Speaker:
     def beep(self, frequency: int = 500, duration: int = 100) -> None: ...
     def play_notes(self, notes: Iterable[str], tempo: int = 120) -> None: ...
-    def play_file(self, file_name: Union[SoundFile, str]) -> None: ...
-    def say(self, text: str) -> None: ...
-    def set_speech_options(
-        self,
-        language: Optional[str] = None,
-        voice: Optional[str] = None,
-        speed: Optional[int] = None,
-        pitch: Optional[int] = None,
-    ): ...
-    def set_volume(self, volume: int, which: str = "_all_") -> None: ...
 
 class Light:
     def on(self, brightness: int = 100) -> None: ...
